Learning/airflow-sample-lecture-main/airflow-sample-lecture-main/requirements.txt
pandas
numpy
requests
apache-airflow-providers-common-sql